requirements.txt
-i https://pypi.org/simple
Flask==2.1.3
Flask-RESTful
Flask-JWT-Extended
Flask-SQLAlchemy
requests
Werkzeug~=2.0.0
gunicorn
psycopg2-binary
uvicorn
emails~=0.6
redmail~=0.4.2
Jinja2~=3.1.2
bcrypt~=4.0.1
pymongo==4.3.3
python-dotenv